% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTable.R
\name{splitTable}
\alias{splitTable}
\title{Table splitter}
\usage{
splitTable(df, maxRows = 74)
}
\arguments{
\item{df}{Data frame to split}

\item{maxRows}{The maximum number of rows for the new table(s)}
}
\value{
A list of tables.
}
\description{
Splits a large table into smaller tables based on a maximum number of rows
}
\examples{
splitTable(airquality, maxRows = 50)
}

