% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRMDKable.R
\name{writeRMDKable}
\alias{writeRMDKable}
\title{rmarkdown Kable Writer}
\usage{
writeRMDKable(df, maxRows = 50, filename = tempfile(),
  slideTitle = "Table", otherText = NULL, repeatOther = TRUE,
  silent = TRUE, ...)
}
\arguments{
\item{df}{A data.frame to split and write out}

\item{maxRows}{The maximum number of rows for the new table(s)}

\item{filename}{The filename of the rmarkdown document including extension.}

\item{slideTitle}{Slide Title}

\item{otherText}{Sub text prior to table}

\item{repeatOther}{Logical. If `TRUE` the sub text will be repeated on each slide}

\item{silent}{Logical. If `TRUE` reporting of the filename is suppressed.}

\item{...}{Other arguments to `kable`}
}
\value{
Nothing in R.  A markdown document with the specified file name
}
\description{
Wrapper to `splitTable` that splits a data frame into a a list and 
writes out markdown slides to be read in as a child document to 
main markdown file
}
\examples{
{
  writeRMDKable(airquality, maxRows = 50, slideTitle = "Slide title", 
                otherText = "Extra information")
  writeRMDKable(airquality, maxRows = 50, slideTitle = "Slide title", 
                otherText = "Extra information",
                col.names = names(airquality))
}
}

