% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTable.R
\name{writeRMDKable}
\alias{writeRMDKable}
\title{rmarkdown Kable Writer}
\usage{
writeRMDKable(df, filename = tempfile(), slideTitle = "Table",
  otherText = NULL, repeatOther = TRUE, ...)
}
\arguments{
\item{df}{A data.frame to write out}

\item{filename}{The filename of the rmarkdown document including extension.}

\item{slideTitle}{Slide Title}

\item{otherText}{Sub text prior to table}

\item{repeatOther}{Logical. If `TRUE` the sub text will be repeated on each slide}

\item{...}{Other arguments to `splitTable`}
}
\value{
Nothing in R.  A markdown documner with the specified file name.
}
\description{
Wrapper to `splitTable` that splits a data frame into a a list and writes out markdown slides to be read in as a child document ot main markdown file
}
\examples{
writeRMDKable(airquality, maxRows = 50, slideTitle = "Slide title", otherText = "Extra information")
}

